pyproject.toml
[project]
name = "django-graph-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "acdh-django-browsing>=3.3.2",
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.6",
    "django-extensions>=4.1",
    "django-spaghetti-and-meatballs>=0.4.2",
    "gunicorn>=25.1.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "lmstudio>=1.5.0",
    "ollama>=0.6.1",
    "pandas>=2.3.2",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.7",
    "ruff>=0.12.12",
]
